[framework, daydream, oculus] move the io package to the gearvrf package
