+++ GVRf/Framework/framework/src/main/java/org/gearvrf/IWearTouchpadEvents.java
@@ -1,4 +1,4 @@
-package org.gearvrf.io;
+package org.gearvrf;
 
 import android.view.MotionEvent;
