+++ GVRf/Framework/framework/src/main/java/org/gearvrf/x3d/AnchorImplementation.java
@@ -27,8 +27,8 @@ import org.gearvrf.GVRRenderData;
 import org.gearvrf.GVRSceneObject;
 import org.gearvrf.ISensorEvents;
 import org.gearvrf.SensorEvent;
-import org.gearvrf.io.GVRControllerType;
-import org.gearvrf.io.GVRInputManager;
+import org.gearvrf.GVRControllerType;
+import org.gearvrf.GVRInputManager;
 import org.gearvrf.scene_objects.GVRCubeSceneObject;
 import org.gearvrf.scene_objects.GVRSphereSceneObject;
 import org.gearvrf.scene_objects.GVRTextViewSceneObject;
