+++ GVRf/Framework/framework/src/main/java/org/gearvrf/x3d/X3Dobject.java
@@ -20,8 +20,8 @@ import android.graphics.Color;
 
 import org.gearvrf.GVRCursorController;
 import org.gearvrf.GVRMeshCollider;
-import org.gearvrf.io.GVRControllerType;
-import org.gearvrf.io.GVRInputManager;
+import org.gearvrf.GVRControllerType;
+import org.gearvrf.GVRInputManager;
 import org.gearvrf.utility.Log;
 
 import java.io.FileNotFoundException;
@@ -75,8 +75,6 @@ import org.gearvrf.scene_objects.GVRSphereSceneObject;
 import org.gearvrf.scene_objects.GVRTextViewSceneObject;
 
 import org.joml.Vector3f;
-import org.joml.AxisAngle4f;
-import org.joml.Matrix4f;
 import org.joml.Quaternionf;
